chore: 쉘 유틸리티 분리 (src/utility.ts)
