+++ src/tool.ts
@@ -1,35 +1,12 @@
 import Database from "bun:sqlite";
-import { execSync } from "node:child_process";
 import { homedir } from "node:os";
 import { join } from "node:path";
 import { tool, ToolDefinition } from "@opencode-ai/plugin";
+import { _, $ } from "./utility";
 
 const DB_PATH = join(homedir(), ".local/share/opencode/opencode.db");
 
 export namespace Tool {
-  function _(command: TemplateStringsArray, ...values: unknown[]) {
-    return command.reduce(
-      (command, segment, index) =>
-        command +
-        segment +
-        (values.length > index
-          ? `'${String(values[index] ?? "")
-              .replace(/\x00/g, "")
-              .replace(/'/g, "'\\''")}'`
-          : ""),
-      "",
-    );
-  }
-  function $(command: string | TemplateStringsArray, ...values: unknown[]) {
-    return execSync(
-      typeof command === "object" && "raw" in command
-        ? _(command, ...values)
-        : command,
-      { encoding: "utf-8" },
-    ).trim();
-  }
-  $.pipe = (...parts: (string | false | null | undefined)[]) =>
-    $(parts.filter(Boolean).join(" "));
   const is = tool.schema;
 
   export abstract class Interface {
